+++ CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - HFP: Emit Audio Connection Released on SLC Release, e.g. remote power off
 - GAP: calculate IO Cap AuthReq Bondable Mode based on `gap_ssp_set_authentication_requirement` and `gap_set_bondable_mode`
 - GAP: only store link key for ssp if remote side has set bondable in io cap auth requirements as well 
+- GAP: allow to disable link supervision timeout
 - HFP HF: only emit single event for RING and AG Status updates
 - HFP AG: fix audio connection setup for in-band ringtone on incoming connection